Move menu classes into their own files.
