--- src/command.py
@@ -1,7 +1,0 @@
-from abc import ABC, abstractmethod
-
-class Command(ABC):
-
-    @abstractmethod
-    def execute(self) -> None:
-        pass
