+++ src/menus/menu.py
@@ -4,7 +4,7 @@
 from abc import ABC, abstractmethod
 from typing import Any
 
-from ui import Prompt, TUI
+from .ui import Prompt, TUI
 
 class BaseMenu(ABC):
     """
